feat:improve Seal API
Moving route to /api/seal. root of this will get all seals by default
and allows for searching, limiting, and offsetting.
Moved both tag and id api endpoint into /tag and /id respectively
removing any overlap between getting a seal with ID or a tag of that
same name, if that were ever to happen