feat:seal creation somewhat working
Storing the Seal in memory rather than commiting right to disk,
as tsoa isn't amazing at saving files.
Using file buffer to generate a md5 hash of the image to use as a
slug.

Errors back when an existing seal is attempted to be created,
failing the @unique constraint for a Seals' slug