feat: it works!
Can go to /seal to get a seal, or /seal/:tag to get a seal with
a specific tag.
Unfortuantely, tsoa does not support sending anything other than
JSON (from the looks of things) so restoring to a normal router
for the image returning endpoints. Might add these manually to
swagger